backend/requirements.txt
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-filter==2.4.0
djoser==2.1.0
requests==2.26.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.10.1
flake8==4.0.1
drf-extra-fields==3.2.1
pillow==8.4.0
